+++ lalstochastic/test/StochasticOptimalFilterTest.c
@@ -132,6 +132,7 @@
  * fprintf()
  * freopen()
  * fabs()
+ * fabsf()
  * \endcode
  *
  * ### Notes ###
@@ -996,8 +997,8 @@ int main(int argc, char *argv[])
     printf("Q(0)=%g + %g i, should be 0\n",
            crealf(optimal.data->data[0]), cimagf(optimal.data->data[0]));
   }
-  if ( fabs(crealf(optimal.data->data[0])) > STOCHASTICOPTIMALFILTERTESTC_TOL
-       || fabs(cimagf(optimal.data->data[0]))
+  if ( fabsf(crealf(optimal.data->data[0])) > STOCHASTICOPTIMALFILTERTESTC_TOL
+       || fabsf(cimagf(optimal.data->data[0]))
        > STOCHASTICOPTIMALFILTERTESTC_TOL )
   {
     printf("  FAIL: Valid data test #1\n");
@@ -1023,10 +1024,10 @@ int main(int argc, char *argv[])
              cimagf(optimal.data->data[i])/crealf(optimal.data->data[1]),
              testNum);
     }
-    if (fabs(crealf(optimal.data->data[i])/crealf(optimal.data->data[1])
+    if (fabsf(crealf(optimal.data->data[i])/crealf(optimal.data->data[1])
              - testNum)/testNum
         > STOCHASTICOPTIMALFILTERTESTC_TOL
-        || fabs(cimagf(optimal.data->data[i])/crealf(optimal.data->data[1]))
+        || fabsf(cimagf(optimal.data->data[i])/crealf(optimal.data->data[1]))
         > STOCHASTICOPTIMALFILTERTESTC_TOL)
     {
       printf("  FAIL: Valid data test #1\n");
@@ -1178,8 +1179,8 @@ int main(int argc, char *argv[])
     printf("Q(0)=%g + %g i, should be 0\n",
            crealf(optimal.data->data[0]), cimagf(optimal.data->data[0]));
   }
-  if ( fabs(crealf(optimal.data->data[0])) > STOCHASTICOPTIMALFILTERTESTC_TOL
-       || fabs(cimagf(optimal.data->data[0]))
+  if ( fabsf(crealf(optimal.data->data[0])) > STOCHASTICOPTIMALFILTERTESTC_TOL
+       || fabsf(cimagf(optimal.data->data[0]))
        > STOCHASTICOPTIMALFILTERTESTC_TOL )
   {
     printf("  FAIL: Valid data test #2\n");
@@ -1202,10 +1203,10 @@ int main(int argc, char *argv[])
              cimagf(optimal.data->data[i])/crealf(optimal.data->data[1]),
              testNum);
     }
-    if (fabs(crealf(optimal.data->data[i])/crealf(optimal.data->data[1])
+    if (fabsf(crealf(optimal.data->data[i])/crealf(optimal.data->data[1])
              - testNum)/testNum
         > STOCHASTICOPTIMALFILTERTESTC_TOL
-        || fabs(cimagf(optimal.data->data[i])/crealf(optimal.data->data[1]))
+        || fabsf(cimagf(optimal.data->data[i])/crealf(optimal.data->data[1]))
         > STOCHASTICOPTIMALFILTERTESTC_TOL)
     {
       printf("  FAIL: Valid data test #2\n");
